+++ sdk/storage/azfile/CHANGELOG.md
@@ -13,6 +13,7 @@
   * Set File/Directory Properties
   * Copy File
 * Added `x-ms-file-last-write-time` request header in Put Range and Put Range from URL.
+* Updated the SAS Version to `2022-11-02` and added `Encryption Scope` to Account SAS.
 
 ### Breaking Changes